chore: seperated posts from feed
