+++ components/shared/index.tsx
@@ -2,5 +2,6 @@ export { default as customButtonStyles } from './customButtonStyles';
 export { default as DashboardFeed } from './Feed';
 export { default as DashboardFooter } from './Footer';
 export { default as DashboardHeader } from './Header';
+export { default as Post } from './IndividualPost';
 export { default as DashboardNavbar } from './Navbar';
 export { default as DashboardSidebar } from './Sidebar';
